MinGW でビルドすると、カラーテーマファイルが読み込めなかったので修正
- swscanf_s() で "%hhu" がうまく扱えなかった
  - BYTE 変数に読み込むために使用
  - __USE_MINGW_ANSI_STDIO=0 でビルドしているのでVSと同じ動作をするはずだった
  - MinGWとmsvcrtなどのバージョンによってうまく動作しない?
- "%u" で unsigned int に読み込むよう修正

+++ teraterm/teraterm/themefile.cpp
@@ -615,9 +615,9 @@ static void LoadColorAttr(const wchar_t *section, const wchar_t *key, const wcha
 		return;
 	}
 
-	BYTE fg_red, fg_green, fg_blue;
-	BYTE bg_red, bg_green, bg_blue;
-	fields = swscanf_s(str, L"%d, %d, %hhu, %hhu, %hhu, %hhu, %hhu, %hhu", &change, &enable, &fg_red, &fg_green,
+	unsigned int fg_red, fg_green, fg_blue;
+	unsigned int bg_red, bg_green, bg_blue;
+	fields = swscanf_s(str, L"%d, %d, %u, %u, %u, %u, %u, %u", &change, &enable, &fg_red, &fg_green,
 					   &fg_blue, &bg_red, &bg_green, &bg_blue);
 	if (fields == 8) {
 		free(str);
